third_party/WebKit/LayoutTests/svg/animations/script-href-no-animate.html

<!DOCTYPE html>
<title>SVGScriptElement.href should not be animatable</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
var foo = 42;

async_test(t => {
  window.onload = () => {
    var s = document.createElementNS('http://www.w3.org/2000/svg', 'script');
    s.onload = t.unreached_func('Should not get a load event');
    s.id = 'x';
    document.querySelector('svg').appendChild(s);
    requestAnimationFrame(_ => {
      requestAnimationFrame(t.step_func_done(_ => {
        assert_equals(foo, 42);
      }));
    });
  };
});
</script>
<svg xmlns:xlink="http://www.w3.org/1999/xlink">
  <set href="#x" attributeName="href" to="resources/set-foo.js"/>
  <set href="#x" attributeName="xlink:href" to="resources/set-foo.js"/>
  <animate href="#x" attributeName="href" to="resources/set-foo.js" dur="0.01s"/>
  <animate href="#x" attributeName="xlink:href" to="resources/set-foo.js" dur="0.01s"/>
</svg>
